Conceptual refactoring.
Data acquisition is split between 4 separate layers of abstraction.

The base layer is the LuaApi wrapper that allows to perform primitive operations to read and write data to and from Lua.
The 2nd layer is QuikLuaApiWrapper that aggregates primitive operations into operational units that work with tables.
The 3rd layer is represented by several QuikApi wrappers that provide trading data of a certain kind.
The 4th layer is a facade that unites all 3rd layer wrappers and provides access to market data to end users